+++ Models/ResidentialMIR.cs
@@ -10,34 +10,34 @@ namespace Alliance_for_Life.Models
         public int Subcontractor { get; set; }
 
         [Required]
-        public int TotBedNights { get; set; }
+        public decimal TotBedNights { get; set; }
 
         [Required]
-        public int TotA2AEnrollment { get; set; }
+        public decimal TotA2AEnrollment { get; set; }
 
         [Required]
-        public int TotA2ABedNights { get; set; }
+        public decimal TotA2ABedNights { get; set; }
 
         [Required]
         public int MonthId { get; set; }
 
         public Month Months { get; set; }
 
-        public double MA2Apercent { get; set; }
+        public decimal MA2Apercent { get; set; }
 
-        public int ClientsJobEduServ { get; set; }
+        public decimal ClientsJobEduServ { get; set; }
 
-        public int ParticipatingFathers { get; set; }
+        public decimal ParticipatingFathers { get; set; }
 
-        public int TotEduClasses { get; set; }
+        public decimal TotEduClasses { get; set; }
 
-        public int TotClientsinEduClasses { get; set; }
+        public decimal TotClientsinEduClasses { get; set; }
 
-        public int TotCaseHrs { get; set; }
+        public decimal TotCaseHrs { get; set; }
 
-        public int TotClientsCaseHrs { get; set; }
+        public decimal TotClientsCaseHrs { get; set; }
 
-        public int TotOtherClasses { get; set; }
+        public decimal TotOtherClasses { get; set; }
 
     }
 }
