+++ ViewModels/ParticipationServicesViewModel.cs
@@ -8,49 +8,49 @@ namespace Alliance_for_Life.ViewModels
     public class ParticipationServicesViewModel
     {
         [Display(Name = "Transportation")]
-        public int PTranspotation { get; set; }
+        public decimal PTranspotation { get; set; }
 
         [Display(Name = "Job Training")]
-        public int PJobTrain { get; set; }
+        public decimal PJobTrain { get; set; }
 
         [Display(Name = "Education Assistance")]
-        public int PEducationAssistance { get; set; }
+        public decimal PEducationAssistance { get; set; }
 
         [Display(Name = "Residential Care")]
-        public int PResidentialCare { get; set; }
+        public decimal PResidentialCare { get; set; }
 
         [Display(Name = "Utilities")]
-        public int PUtilities { get; set; }
+        public decimal PUtilities { get; set; }
 
         [Display(Name = "Housing Emergency")]
-        public int PHousingEmergency { get; set; }
+        public decimal PHousingEmergency { get; set; }
 
         [Display(Name = "Housing Assistance")]
-        public int PHousingAssistance { get; set; }
+        public decimal PHousingAssistance { get; set; }
 
         [Display(Name = "Child Care")]
-        public int PChildCare { get; set; }
+        public decimal PChildCare { get; set; }
 
         [Display(Name = "Clothing")]
-        public int PClothing { get; set; }
+        public decimal PClothing { get; set; }
 
         [Display(Name = "Food")]
-        public int PFood { get; set; }
+        public decimal PFood { get; set; }
 
         [Display(Name = "Supplies")]
-        public int PSupplies { get; set; }
+        public decimal PSupplies { get; set; }
 
         [Display(Name = "Other")]
-        public int POther { get; set; }
+        public decimal POther { get; set; }
 
         [Display(Name = "Other")]
-        public int POther2 { get; set; }
+        public decimal POther2 { get; set; }
 
         [Display(Name = "Other")]
-        public int POther3 { get; set; }
+        public decimal POther3 { get; set; }
 
         [Display(Name = "Participation Totals")]
-        public int PTotals { get; set; }
+        public decimal PTotals { get; set; }
 
         public int Id { get; internal set; }
